add support for N25Q256
I needed to reflash a N25Q256 on a host system and hacked flashrom
to make this work. For this, I used the 4-Bytes address support
developed by Boris Baykov. This is why the patch is an RFC. The
whole patchset is available under github here:

	https://github.com/legoater/flashrom

I can help in getting this patchset merged but do the comments
in :

	http://patchwork.coreboot.org/patch/4272/

need to be all addressed or can we rework a bit the patchset to
merge the main changes separately :

       - 4-bytes addressing
       - W25Q256.V support
       - SFDP
       - Progress visualization
       - ...

Tested under Linux with spidev.

Signed-off-by: Cédric Le Goater <clg@kaod.org>